requirements.txt
modal
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography] 
passlib[bcrypt]
python-multipart